feature - import_export and resuts modularized: - two additional blueprints are created, because import-export should not be stcuk in result wrapping
