+++ assets/3_styles_small_desktop.css
@@ -67,17 +67,17 @@
 
     /* The import-export section has index 1 */
     /* We target the extreme right mini-box(s) */
-    .mini-box:nth-of-type(3), .mini-box:nth-of-type(6) {
+    .mini-box:nth-of-type(2), .mini-box:nth-of-type(5) {
         margin-right: 0;
     }    
 
     /* We target the top row of mini-box(s) */
-    .mini-box:nth-of-type(2), .mini-box:nth-of-type(3) {
+    .mini-box:nth-of-type(1), .mini-box:nth-of-type(2) {
         width: calc((50%) - (6px));
     }    
 
     /* We target the second row of mini-boxe(s) */
-    .mini-box:nth-of-type(n+4) {
+    .mini-box:nth-of-type(n+3) {
         width: calc((33%) - (8px));
         flex-direction: column;
         height: 260px;
