+++ assets/4_styles_large_desktop.css
@@ -61,7 +61,7 @@
     ---------------------------------- MINI-BOX ----------------------------------------
     */
 
-    .mini-box:nth-of-type(n+4) {
+    .mini-box:nth-of-type(n+3) {
         height: 240px;
     }
